+++ src/test/java/com/petit/toon/controller/RestDocsSupport.java
@@ -1,6 +1,6 @@
 package com.petit.toon.controller;
 
-import com.petit.toon.util.JwtUtil;
+import com.petit.toon.security.JwtTokenProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -23,7 +23,7 @@ public abstract class RestDocsSupport {
     protected MockMvc mockMvc;
 
     @MockBean
-    private JwtUtil jwtUtil;
+    private JwtTokenProvider jwtTokenProvider;
 
     @BeforeEach
     void setUp(WebApplicationContext context,
